# shared/openapi.yaml
openapi: 3.1.0
info:
  title: VOCPage API
  version: 0.0.1
  description: VOC (Voice of Customer) management system REST API
servers:
  - url: /api
    description: Local dev
...
paths:
...
  /vocs:
    get:
      tags: [vocs]
      operationId: listVocs
      summary: List VOCs (paginated)
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VocStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/VocPriority'
        - name: systemId
          in: query
          schema:
            type: string
            format: uuid
        - name: menuId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
            format: uuid
        - name: assigneeId
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: includeDeleted
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated VOC list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVocs'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [vocs]
      operationId: createVoc
      summary: Create VOC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voc'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /vocs/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [vocs]
      operationId: getVoc
      summary: Get VOC detail
      responses:
        '200':
          description: VOC
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voc'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [vocs]
      operationId: updateVoc
      summary: Update VOC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocPatch'
      responses:
        '200':
          description: Updated VOC
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voc'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [vocs]
      operationId: deleteVoc
      summary: Soft-delete VOC (admin only)
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vocs/{id}/status:
    parameters:
      - $ref: '#/components/parameters/id'
    patch:
      tags: [vocs]
      operationId: updateVocStatus
      summary: Update VOC status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/VocStatus'
      responses:
        '200':
          description: Updated VOC
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voc'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /notifications:
    get:
      tags: [notifications]
      operationId: listNotifications
      summary: List notifications for current user
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/unread-count:
    get:
      tags: [notifications]
      operationId: getUnreadCount
      summary: Get unread notification count
      responses:
        '200':
          description: Count
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/id'
    patch:
      tags: [notifications]
      operationId: markNotificationRead
      summary: Mark notification as read
      responses:
        '200':
          description: Marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    VocStatus:
      type: string
      enum: ['접수', '검토중', '처리중', '완료', '드랍']
    VocPriority:
      type: string
      enum: [urgent, high, medium, low]
    VocSource:
      type: string
      enum: [manual, import]
    Voc:
      type: object
      required: [id, issue_code, sequence_no, title, status, priority, source, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        issue_code:
          type: string
        sequence_no:
          type: integer
        title:
          type: string
        body:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/VocStatus'
        priority:
          $ref: '#/components/schemas/VocPriority'
        voc_type_id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
          nullable: true
        menu_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
          nullable: true
        author_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        source:
          $ref: '#/components/schemas/VocSource'
        review_status:
          type: string
          nullable: true
        structured_payload:
          type: object
          additionalProperties: true
          nullable: true
        resolution_quality:
          type: string
          enum: ['근본해결', '임시조치']
          nullable: true
        drop_reason:
          type: string
          enum: ['중복', '정책거부', '재현불가', '범위외', '기타']
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
    VocInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        body:
          type: string
        status:
          $ref: '#/components/schemas/VocStatus'
        priority:
          $ref: '#/components/schemas/VocPriority'
        voc_type_id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        menu_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
        source:
          $ref: '#/components/schemas/VocSource'
    VocPatch:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        status:
          $ref: '#/components/schemas/VocStatus'
        priority:
          $ref: '#/components/schemas/VocPriority'
        voc_type_id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        menu_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        review_status:
          type: string
          enum: ['unverified', 'approved', 'rejected', 'pending_deletion']
        resolution_quality:
          type: string
          enum: ['근본해결', '임시조치']
        drop_reason:
          type: string
          enum: ['중복', '정책거부', '재현불가', '범위외', '기타']
...
    Notification:
      type: object
      required: [id, user_id, type, message, is_read, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: ['comment', 'status_change', 'assigned']
        voc_id:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
        is_read:
          type: boolean
        voc_priority:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    PaginatedVocs:
      type: object
      required: [items, total, page, limit, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Voc'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...